+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Release Date: TBD
   * Probability.sample_without_replacement (alias: resample_without_replacement, jackknife)
   * Frequencies.cumulative_distribution_function (alias: cdf, cumulative_distribution)
   * Frequencies.cumulative_distribution_function_value (alias: cdf_value, cumulative_distribution_value)
+  * Array.sample_with_replacement (alias: resample_with_replacement, bootstrap)
+  * Array.sample_without_replacement (alias: resample_without_replacement, jackknife)
   * Collection.collect
   * Collection.catalog (alias: catalogue)
   * Collection.catalog_hash(alias: catalogue_hash)